vin-regioner and vin-producenter to regioner and producenter and removed redirection
