+++ next.config.mjs
@@ -23,32 +23,6 @@ const nextConfig = {
       },
     ],
   },
-  async redirects() {
-    return [
-      // for vin producenter
-      {
-        source: '/producenter',
-        destination: '/vin-producenter',
-        permanent: true,
-      },
-      {
-        source: '/vin-producenter/page',
-        destination: '/vin-producenter',
-        permanent: true,
-      },
-      {
-        source: '/vin-producenter/page/1',
-        destination: '/vin-producenter',
-        permanent: true,
-      },
-      // for vin regioner
-      {
-        source: '/regioner',
-        destination: '/vin-regioner',
-        permanent: true,
-      },
-    ];
-  },
 };
 
 export default nextConfig;
