+++ src/app/producenter/page.jsx
@@ -1,0 +1,31 @@
+import { countProducenters } from '@/src/lib/api/producenterAPI';
+import BreadCrumb from '../Components/breadcrumb/BreadCrumb';
+import ProducenterContainer from './components/ProducenterContainer';
+import SubscriptionForm from '../Components/subscription/SubscriptionForm';
+import CatAccordion from '../[category]/Components/CatAccordion';
+import Sidebar from '../Components/Sidebar';
+
+export default async function page() {
+  const totalProducenters = await countProducenters();
+  return (
+    <>
+      <div className="container mx-auto flex  lg:flex-row gap-6 lg:gap-12 mt-4 lg:mt-10 bg-slate-50  p-4">
+        <div className="md:w-2/3 ">
+          <div className=" text-xs lg:text-sm flex gap-1 my-2">
+            <BreadCrumb title1="Producenter" />
+          </div>
+          <h1 className="text-2xl lg:text-3xl  mb-4 font-bold uppercase">Producenter</h1>
+          <ProducenterContainer totalProducenters={totalProducenters} />
+          <div className="space-y-10 my-10">
+            <SubscriptionForm />
+            <CatAccordion />
+          </div>
+        </div>
+
+        <div className="w-1/3 hidden md:block sticky top-0 h-full">
+          <Sidebar />
+        </div>
+      </div>
+    </>
+  );
+}
